tasks_manager_fast_api/requirements.txt
fastapi==0.111.0
sqlalchemy==2.0.30
python-dotenv==1.0.1
psycopg2-binary==2.9.9
alembic==1.13.1
pydantic==2.7.2